requirements.txt
beautifulsoup4
mako
pyphen
pyyaml
requests
